requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
bcrypt==4.2.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
Django==4.2
django-celery-results==2.5.1
django-cors-headers==4.5.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
dotmap==1.3.30
ffmpeg-progress-yield==0.9.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
kombu==5.4.2
mutagen==1.47.0
paramiko==3.5.0
pillow==11.0.0
prompt_toolkit==3.0.48
pyaes==1.6.1
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.22
pycryptodomex==3.21.0
pyftpdlib==2.0.1
PyNaCl==1.5.0
Pyrogram @ https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.30.zip#sha256=d177c7fe42c830879070614c6c7d42536d62c54506200f5dbdd7b1fbb7c1d24d
pyromod==3.1.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.2
redis==5.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
TgCrypto==1.2.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
wheel==0.44.0
yt-dlp==2024.10.22
